+++ include/iot/data.h
@@ -2055,6 +2055,16 @@ extern const iot_data_t * iot_data_vector_find (const iot_data_t * vector, iot_d
 extern void iot_data_iter (const iot_data_t * iterable, iot_data_iter_t * iter);
 
 /**
+ * @brief Get the number of elements in the underlying iterable data of an iterator
+ *
+ * The function returns the number of elements in the iterable data of an iterator, when the iterable data is of type list, map or vector
+ *
+ * @param iter Input iterator
+ * @return Number of elements in underlying interable data
+ */
+extern uint32_t iot_data_iter_size (const iot_data_iter_t * iter);
+
+/**
  * @brief Iterate to next iterable element
  *
  * The function to set the iterator to point to the next element in an iterable. On reaching the end of the iterable,