+++ RELEASE_NOTES.md
@@ -379,3 +379,4 @@
 
 - Support added for Debian 12
 - Added `iot_data_iter_t` to allow for iteration over generic iterable types
+- Added `iot_data_iter_size` function for determining number of elements for generic iterator